+++ stmt_test.go
@@ -43,7 +43,7 @@ func TestBoltStmt_SelectOne(t *testing.T) {
 		t.Fatalf("An error occurred getting next row: %s", err)
 	}
 
-	if output[0].(int8) != 1 {
+	if output[0].(int64) != 1 {
 		t.Fatalf("Unexpected output. Expected 1. Got: %d", output)
 	}
 
@@ -89,7 +89,7 @@ func TestBoltStmt_SelectMany(t *testing.T) {
 		t.Fatalf("An error occurred getting next row: %s", err)
 	}
 
-	if output[0].(int8) != 1 {
+	if output[0].(int64) != 1 {
 		t.Fatalf("Unexpected output. Expected 1. Got: %#v", output[0])
 	}
 	if output[1].(float64) != 34234.34323 {
@@ -98,7 +98,7 @@ func TestBoltStmt_SelectMany(t *testing.T) {
 	if output[2].(string) != "string" {
 		t.Fatalf("Unexpected output. Expected string. Got: %#v", output[2])
 	}
-	if !reflect.DeepEqual(output[3].([]interface{}), []interface{}{int8(1), "2", int8(3), true, interface{}(nil)}) {
+	if !reflect.DeepEqual(output[3].([]interface{}), []interface{}{int64(1), "2", int64(3), true, interface{}(nil)}) {
 		t.Fatalf("Unexpected output. Expected []interface{}{1, '2', 3, true, nil}. Got: %#v", output[3])
 	}
 	if !output[4].(bool) {
@@ -137,7 +137,7 @@ func TestBoltStmt_InvalidArgs(t *testing.T) {
 		"a": 1,
 		"b": 34234.34323,
 		"c": "string",
-		"d": []interface{}{int8(1), "2", int8(3), true, interface{}(nil)},
+		"d": []interface{}{int64(1), "2", int64(3), true, interface{}(nil)},
 		"e": true,
 		"f": nil,
 	}
@@ -269,7 +269,7 @@ func TestBoltStmt_CreateArgs(t *testing.T) {
 		t.Fatalf("An error occurred getting next row: %s", err)
 	}
 
-	if output[0].(int8) != 1 {
+	if output[0].(int64) != 1 {
 		t.Fatalf("Unexpected output. Expected 1. Got: %#v", output[0])
 	}
 	if output[1].(float64) != 34234.34323 {
@@ -278,7 +278,7 @@ func TestBoltStmt_CreateArgs(t *testing.T) {
 	if output[2].(string) != "string" {
 		t.Fatalf("Unexpected output. Expected string. Got: %#v", output[2])
 	}
-	if !reflect.DeepEqual(output[3].([]interface{}), []interface{}{int8(1), int8(2), int8(3)}) {
+	if !reflect.DeepEqual(output[3].([]interface{}), []interface{}{int64(1), int64(2), int64(3)}) {
 		t.Fatalf("Unexpected output. Expected []interface{}{1, 2, 3}. Got: %#v", output[3])
 	}
 	if !output[4].(bool) {
@@ -639,7 +639,7 @@ func TestBoltStmt_SingleRel(t *testing.T) {
 		"b": int64(math.MaxInt64),
 		"c": true,
 		"d": nil,
-		"e": []interface{}{int8(1), int8(2), int8(3)},
+		"e": []interface{}{int64(1), int64(2), int64(3)},
 		"f": 3.4,
 		"g": int32(math.MaxInt32),
 		"h": false,
@@ -666,13 +666,13 @@ func TestBoltStmt_SingleRel(t *testing.T) {
 	if output[0].(graph.Relationship).Properties["d"] != nil {
 		t.Fatalf("Unexpected return data: %#v", output)
 	}
-	if !reflect.DeepEqual(output[0].(graph.Relationship).Properties["e"].([]interface{}), []interface{}{int8(1), int8(2), int8(3)}) {
+	if !reflect.DeepEqual(output[0].(graph.Relationship).Properties["e"].([]interface{}), []interface{}{int64(1), int64(2), int64(3)}) {
 		t.Fatalf("Unexpected return data: %#v", output)
 	}
 	if output[0].(graph.Relationship).Properties["f"].(float64) != 3.4 {
 		t.Fatalf("Unexpected return data: %#v", output)
 	}
-	if output[0].(graph.Relationship).Properties["g"].(int32) != math.MaxInt32 {
+	if output[0].(graph.Relationship).Properties["g"].(int64) != math.MaxInt32 {
 		t.Fatalf("Unexpected return data: %#v", output)
 	}
 	if output[0].(graph.Relationship).Properties["h"].(bool) {
@@ -714,7 +714,7 @@ func TestBoltStmt_SingleNode(t *testing.T) {
 		"b": 1,
 		"c": true,
 		"d": nil,
-		"e": []interface{}{int8(1), int8(2), int8(3)},
+		"e": []interface{}{int64(1), int64(2), int64(3)},
 		"f": 3.4,
 		"g": -1,
 		"h": false,
@@ -732,7 +732,7 @@ func TestBoltStmt_SingleNode(t *testing.T) {
 	if output[0].(graph.Node).Properties["a"].(string) != "foo" {
 		t.Fatalf("Unexpected return data: %#v", output)
 	}
-	if output[0].(graph.Node).Properties["b"].(int8) != 1 {
+	if output[0].(graph.Node).Properties["b"].(int64) != 1 {
 		t.Fatalf("Unexpected return data: %#v", output)
 	}
 	if !output[0].(graph.Node).Properties["c"].(bool) {
@@ -741,13 +741,13 @@ func TestBoltStmt_SingleNode(t *testing.T) {
 	if output[0].(graph.Node).Properties["d"] != nil {
 		t.Fatalf("Unexpected return data: %#v", output)
 	}
-	if !reflect.DeepEqual(output[0].(graph.Node).Properties["e"].([]interface{}), []interface{}{int8(1), int8(2), int8(3)}) {
+	if !reflect.DeepEqual(output[0].(graph.Node).Properties["e"].([]interface{}), []interface{}{int64(1), int64(2), int64(3)}) {
 		t.Fatalf("Unexpected return data: %#v", output)
 	}
 	if output[0].(graph.Node).Properties["f"].(float64) != 3.4 {
 		t.Fatalf("Unexpected return data: %#v", output)
 	}
-	if output[0].(graph.Node).Properties["g"].(int8) != -1 {
+	if output[0].(graph.Node).Properties["g"].(int64) != -1 {
 		t.Fatalf("Unexpected return data: %#v", output)
 	}
 	if output[0].(graph.Node).Properties["h"].(bool) {
@@ -808,25 +808,25 @@ func TestBoltStmt_SelectIntLimits(t *testing.T) {
 	if output[0].(int64) != math.MinInt64 {
 		t.Fatalf("Unexpected output. Expected %d. Got: %d", math.MinInt64, output[0])
 	}
-	if output[1].(int32) != math.MinInt32 {
+	if output[1].(int64) != math.MinInt32 {
 		t.Fatalf("Unexpected output. Expected %d. Got: %d", math.MinInt32, output[1])
 	}
-	if output[2].(int16) != math.MinInt16 {
+	if output[2].(int64) != math.MinInt16 {
 		t.Fatalf("Unexpected output. Expected %d. Got: %d", math.MinInt16, output[2])
 	}
-	if output[3].(int8) != math.MinInt8 {
+	if output[3].(int64) != math.MinInt8 {
 		t.Fatalf("Unexpected output. Expected %d. Got: %d", math.MinInt8, output[3])
 	}
-	if output[4].(int8) != -16 {
+	if output[4].(int64) != -16 {
 		t.Fatalf("Unexpected output. Expected -16. Got: %d", output[4])
 	}
-	if output[5].(int8) != math.MaxInt8 {
+	if output[5].(int64) != math.MaxInt8 {
 		t.Fatalf("Unexpected output. Expected %d. Got: %d", math.MaxInt8, output[5])
 	}
-	if output[6].(int16) != math.MaxInt16 {
+	if output[6].(int64) != math.MaxInt16 {
 		t.Fatalf("Unexpected output. Expected %d. Got: %d", math.MaxInt16, output[6])
 	}
-	if output[7].(int32) != math.MaxInt32 {
+	if output[7].(int64) != math.MaxInt32 {
 		t.Fatalf("Unexpected output. Expected %d. Got: %d", math.MaxInt32, output[7])
 	}
 	if output[8].(int64) != math.MaxInt64 {
@@ -1168,13 +1168,13 @@ func TestBoltStmt_PipelineExec(t *testing.T) {
 		t.Fatalf("An error occurred on getting row: %s", err)
 	}
 
-	if output[0].(graph.Node).Labels[0] != "FOO" && output[0].(graph.Node).Properties["b"] != int8(1) && output[0].(graph.Node).Properties["a"] != "two" {
+	if output[0].(graph.Node).Labels[0] != "FOO" && output[0].(graph.Node).Properties["b"] != int64(1) && output[0].(graph.Node).Properties["a"] != "two" {
 		t.Fatalf("Unexpected return data: %s", err)
 	}
-	if output[1].(graph.Node).Labels[0] != "BAR" && output[1].(graph.Node).Properties["a"] != int8(2) && output[1].(graph.Node).Properties["b"] != "two" {
+	if output[1].(graph.Node).Labels[0] != "BAR" && output[1].(graph.Node).Properties["a"] != int64(2) && output[1].(graph.Node).Properties["b"] != "two" {
 		t.Fatalf("Unexpected return data: %s", err)
 	}
-	if output[2].(graph.Node).Labels[0] != "BAZ" && output[2].(graph.Node).Properties["a"] != int8(3) && output[2].(graph.Node).Properties["b"] != "four" {
+	if output[2].(graph.Node).Labels[0] != "BAZ" && output[2].(graph.Node).Properties["a"] != int64(3) && output[2].(graph.Node).Properties["b"] != "four" {
 		t.Fatalf("Unexpected return data: %s", err)
 	}
 
@@ -1275,13 +1275,13 @@ func TestBoltStmt_PipelineQuery(t *testing.T) {
 		t.Fatalf("Error getting final rows: %s", err)
 	}
 
-	if foo[0].(graph.Node).Labels[0] != "FOO" && foo[0].(graph.Node).Properties["b"] != int8(1) && foo[0].(graph.Node).Properties["a"] != "two" {
+	if foo[0].(graph.Node).Labels[0] != "FOO" && foo[0].(graph.Node).Properties["b"] != int64(1) && foo[0].(graph.Node).Properties["a"] != "two" {
 		t.Fatalf("Unexpected return data: %s", err)
 	}
-	if bar[0].(graph.Node).Labels[0] != "BAR" && bar[0].(graph.Node).Properties["a"] != int8(2) && bar[0].(graph.Node).Properties["b"] != "two" {
+	if bar[0].(graph.Node).Labels[0] != "BAR" && bar[0].(graph.Node).Properties["a"] != int64(2) && bar[0].(graph.Node).Properties["b"] != "two" {
 		t.Fatalf("Unexpected return data: %s", err)
 	}
-	if baz[0].(graph.Node).Labels[0] != "BAZ" && baz[0].(graph.Node).Properties["a"] != int8(3) && baz[0].(graph.Node).Properties["b"] != "four" {
+	if baz[0].(graph.Node).Labels[0] != "BAZ" && baz[0].(graph.Node).Properties["a"] != int64(3) && baz[0].(graph.Node).Properties["b"] != "four" {
 		t.Fatalf("Unexpected return data: %s", err)
 	}
 	if nextRows != nil {
@@ -1447,7 +1447,7 @@ func TestBoltStmt_PipelineQueryCloseMiddle(t *testing.T) {
 		t.Fatalf("Error closing pipeline rows: %s", err)
 	}
 
-	if foo[0].(graph.Node).Labels[0] != "FOO" && foo[0].(graph.Node).Properties["b"] != int8(1) && foo[0].(graph.Node).Properties["a"] != "two" {
+	if foo[0].(graph.Node).Labels[0] != "FOO" && foo[0].(graph.Node).Properties["b"] != int64(1) && foo[0].(graph.Node).Properties["a"] != "two" {
 		t.Fatalf("Unexpected return data: %s", err)
 	}
 
@@ -1542,7 +1542,7 @@ func TestBoltStmt_SqlQueryAndExec(t *testing.T) {
 	if err != nil {
 		t.Fatalf("Error occurred decoding item: %s", err)
 	}
-	if !reflect.DeepEqual(dVal.([]interface{}), []interface{}{int8(1), int8(2), int8(3)}) {
+	if !reflect.DeepEqual(dVal.([]interface{}), []interface{}{int64(1), int64(2), int64(3)}) {
 		t.Fatalf("Unexpected value for d. Expected: %#v  Got: %#v", []interface{}{1, 2, 3}, dVal)
 	}
 
@@ -1561,8 +1561,8 @@ func TestBoltStmt_SqlQueryAndExec(t *testing.T) {
 	if nodeVal.(graph.Node).Labels[0] != "FOO" {
 		t.Fatalf("Unexpected label for node. Expected: %#v  Got: %#v", "FOO", nodeVal)
 	}
-	if nodeVal.(graph.Node).Properties["a"] != int8(1) {
-		t.Fatalf("Unexpected value for node. Expected: %#v  Got: %#v", int8(1), nodeVal)
+	if nodeVal.(graph.Node).Properties["a"] != int64(1) {
+		t.Fatalf("Unexpected value for node. Expected: %#v  Got: %#v", int64(1), nodeVal)
 	}
 
 	relVal, err := encoding.Unmarshal(rel)
